+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fix zigbee shutter property access
+
 ## 3.2.10-alpha.0 (2025-07-03)
 
 * (theimo1221) Centralise some shutter functions